pyproject.toml
[tool.poetry]
name = "firefly"
version = "5.40.0"
description = "Desktop client application for Nebula - Open source media asset management and broadcast automation system."
authors = ["Martin Wacker <martas@imm.cz>"]

[tool.poetry.dependencies]
python = "^3.8"
PyQt5 = "^5.15.6"
websocket-client = "^1.3.1"
nxtools = "^1.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
